tb-social-media: Ollama-Provider + LLM-Dispatcher (Slice E2b)
Port von llm/ollama.py + dispatcher.py. Default-Pfad: lokales Ollama.

- LlmProvider-Trait (async-trait: generate + generate_text).
- OllamaProvider: POST <host>/api/generate (stream=false, options
  temperature/num_predict/top_p; format=json wenn System-Prompt strict-json
  verlangt), from_env (OLLAMA_MODEL/HOST def 127.0.0.1:11434, host-Normalisierung
  + ://-Default), generate → render_user_prompt → generate_text → parse_llm_payload
  (cost 0.0). Connect-Fehler → ProviderUnavailable, Timeout/HTTP → ProviderError,
  leere Antwort → Fehler.
- LlmDispatcher: resolve_consent (external_llm_consent / Override),
  resolve_provider_name (Override/SOCIAL_MEDIA_LLM_PROVIDER/ollama),
  Consent-Gate (externe Provider ohne Consent → Ollama), Fallback-Chain
  (primär → ollama). minimax/claude_haiku noch nicht gebaut → Unavailable →
  Fallback (E2c). cost_usd_estimate i64→f64 korrigiert (LlmResponse PartialEq).

async-trait-Dep. 4 neue Tests (Ollama generate via wiremock, leere Antwort
→ Fehler, candidate-chain, host-Normalisierung). tb-social-media-Tests gruen.

+++ rust/crates/tb-social-media/src/llm.rs
@@ -43,14 +43,14 @@ pub struct LlmRequest {
 }
 
 /// Plattform-spezifische LLM-Antwort.
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct LlmResponse {
     pub youtube: PlatformEnrichment,
     pub tiktok: PlatformEnrichment,
     pub instagram: PlatformEnrichment,
     pub provider: String,
     pub model: String,
-    pub cost_usd_estimate: Option<i64>,
+    pub cost_usd_estimate: Option<f64>,
 }
 
 /// Freitext-Antwort (für generate_text-Provider).
@@ -260,7 +260,7 @@ fn find_json_object(text: &str) -> Result<String, LlmError> {
 }
 
 /// Parst die Roh-LLM-Ausgabe in eine [`LlmResponse`] (Python `parse_llm_payload`).
-pub fn parse_llm_payload(raw_text: &str, provider: &str, model: &str, cost: Option<i64>) -> Result<LlmResponse, LlmError> {
+pub fn parse_llm_payload(raw_text: &str, provider: &str, model: &str, cost: Option<f64>) -> Result<LlmResponse, LlmError> {
     let json_text = find_json_object(raw_text)?;
     let payload: Value = serde_json::from_str(&json_text)
         .map_err(|e| LlmError::ProviderError(format!("invalid JSON from LLM: {e}")))?;
